Replace termios crate with libc definitions
termios(3) definitions were added to libc in v0.2.3. Baud rate constants
were added to libc in v0.2.21.

Utilitizing the libc definitions for the termios API enables dropping
the termios crate as a dependency. Fewer dependencies is nice in its own
right, but this also paves the way for a simpler implementation of
custom baud rates on Linux (#49).

Custom baud rates on Linux are best implemented with the termios2 API
provided by the Linux kernel, which is not part of the termios crate.
This most likely means the Linux implementation will use the "unsafe"
interfaces directly for termios2. Doing the same for other termios-based
implementations allows the implementations to remain as similar as
possible, which makes it easier to maintain both implementations in
parallel.