Replace ioctl-rs crate with libc definitions
Constants from tty_ioctl(4) were added to the libc crate in v0.2.19. In
particular, TIOCEXCL, TIOCNXCL, TIOCMBIS, TIOCMBIC, and TIOCMGET are
relevant to this crate. These definitions were previously provided by
the ioctl-rs crate. As the diffstat for this commit shows, the remaining
utility of ioctl-rs (providing a safe wrapper around ioctl) isn't enough
of a value-add to warrant keeping it as a dependency.

Less is more.