Remove superfluous memory allocation
Every time read() or write() is called, an array of pollfd structs was
allocated on the heap with the vec!() macro; however, it's always a
single pollfd. This can just as easily be allocated on the stack to
avoid an extra malloc() and free() for every call to poll()/ppoll().