Add license header to source files
Due to some confusion about the project's license, I've decided to make
the license more prominent by adding a license header to all of the
source files. Several people seem to think that they're entitled to
commercial-level support when the license clearly states "PROVIDED AS
IS" and "WITHOUT WARRANTY." Yet others seem to think that the MIT
license is synonymous with being in the public domain, because they've
managed to plagiarize my work. Literally the only condition of the
license is to not plagiarize:

> The above copyright notice and this permission notice shall be
> included in all copies or substantial portions of the Software.

Neither of these behaviors is acceptable.

Somehow, I doubt many people from the entitled group bother to read the
source code (or LICENSE for that matter), so this probably won't help
much with that issue, but at least this makes it more annoying to
plagiarize.